+++ src/features/home/components/fib-spiral/fib-spiral.component.tsx
@@ -37,18 +37,21 @@ export const FibSpiral: React.FC<Props> = ({ seconds }) => {
   }, [minutes])
 
   return (
-    <Wrapper offset={START_SPIRAL_OFFSET + progress} color={fibColor}>
+    <Wrapper
+      className="progress-spiral"
+      offset={START_SPIRAL_OFFSET + progress}
+      color={fibColor}
+    >
       <svg
         width="100%"
         height="100%"
-        className="progress-spiral"
         xmlns="http://www.w3.org/2000/svg"
         fill="none"
         viewBox="0 0 760 769"
       >
         <path
           stroke="var(--c-spiral)"
-          className="progress-trial"
+          className="spiral-trial"
           strokeLinecap="round"
           strokeLinejoin="round"
           strokeWidth="1rem"
@@ -57,9 +60,8 @@ export const FibSpiral: React.FC<Props> = ({ seconds }) => {
         />
 
         <path
-          strokeDasharray={MAX_SPIRAL_VALUE}
           strokeDashoffset={START_SPIRAL_OFFSET}
-          className="progress-path"
+          className="spiral-path"
           stroke="currentColor"
           strokeLinecap="round"
           strokeLinejoin="round"