+++ src/features/home/components/fib-spiral/fib-spiral.constants.ts
@@ -1,5 +1,6 @@
-export const START_SPIRAL_OFFSET = -1285
+export const START_SPIRAL_OFFSET = 0
 export const MAX_SPIRAL_VALUE = 1285
+export const INIT_STROKE_DASHARRAY = `0 ${MAX_SPIRAL_VALUE} ${MAX_SPIRAL_VALUE}`
 
 export const PATH_TO_DRAWN = `
   M 380.5.8